refactor #16 : Header 컴포넌트 분리
